move existing stores to own module directory
