+++ frontend/src/store/modules/ux/ux.js
@@ -5,14 +5,14 @@ import {
     TableIcon, TemplateIcon, UserGroupIcon, UsersIcon
 } from '@heroicons/vue/outline'
 
-import DeviceGroupOutlineIcon from '../components/icons/DeviceGroupOutline.js'
-import PipelinesIcon from '../components/icons/Pipelines.js'
-import ProjectsIcon from '../components/icons/Projects.js'
-import usePermissions from '../composables/Permissions.js'
-import { Roles } from '../utils/roles.js'
+import DeviceGroupOutlineIcon from '../../../components/icons/DeviceGroupOutline.js'
+import PipelinesIcon from '../../../components/icons/Pipelines.js'
+import ProjectsIcon from '../../../components/icons/Projects.js'
+import usePermissions from '../../../composables/Permissions.js'
+import { Roles } from '../../../utils/roles.js'
 
-import commonActions from './common/actions.js'
-import commonMutations from './common/mutations.js'
+import commonActions from '../../common/actions.js'
+import commonMutations from '../../common/mutations.js'
 
 const initialState = () => ({
     leftDrawer: {
